+++ lib/win32/pcre/pcre_scanner_unittest.cc
@@ -37,6 +37,7 @@
 #endif
 
 #include <stdio.h>
+#include <string.h>      /* for strchr */
 #include <string>
 #include <vector>
 
@@ -106,11 +107,11 @@ static void TestScanner() {
   CHECK_EQ(comments[2].as_string(), " /* and here is gamma */\n");
   comments.resize(0);
 
-  s.GetComments(0, strchr(input, '/') - input, &comments);
+  s.GetComments(0, (int)(strchr(input, '/') - input), &comments);
   CHECK_EQ(comments.size(), 0);
   comments.resize(0);
 
-  s.GetComments(strchr(input, '/') - input - 1, sizeof(input),
+  s.GetComments((int)(strchr(input, '/') - input - 1), sizeof(input),
                 &comments);
   CHECK_EQ(comments.size(), 3);
   CHECK_EQ(comments[0].as_string(), " // this sets alpha\n");
@@ -118,8 +119,8 @@ static void TestScanner() {
   CHECK_EQ(comments[2].as_string(), " /* and here is gamma */\n");
   comments.resize(0);
 
-  s.GetComments(strchr(input, '/') - input - 1,
-                strchr(input + 1, '\n') - input + 1, &comments);
+  s.GetComments((int)(strchr(input, '/') - input - 1),
+                (int)(strchr(input + 1, '\n') - input + 1), &comments);
   CHECK_EQ(comments.size(), 1);
   CHECK_EQ(comments[0].as_string(), " // this sets alpha\n");
   comments.resize(0);
